[fixup] Projection feedback
